+++ CHANGELOG.md
@@ -16,6 +16,12 @@ and this project adheres to
 - [@http/interceptor/cors] `cors()` now returns an `Interceptors` objects for
   direct use with `intercept()` or `init()`.
 
+### Added
+
+- [@http/response] `setHeaders()`, `headerEntries()` to accompany
+  `appendHeaders()`
+- [@http/request] `memoize()`
+
 ## [0.7.0]
 
 ### Added